refactor: font info moved to its own package
